fix: improve user admin checks in vehicle service functions

+++ microservices/vehicle/services/vehicle_service.py
@@ -5,8 +5,11 @@ from typing import List
 
 def create_vehicle_service(user_id : int, vehicle_model: VehicleModel):
     is_user_admin = select_user_is_admin(user_id)
-    if is_user_admin[0]["is_admin"]:
-        return insert_vehicle(vehicle_model)
+    if is_user_admin:
+        if is_user_admin[0]["is_admin"]:
+            return insert_vehicle(vehicle_model)
+    else:
+            return { "message" : "No such user"}
     return { "message" : "Access Denied"}
 
 def get_all_vehicles_service():
@@ -17,12 +20,18 @@ def get_vehicle_by_id_service(vehicle_id: int):
 
 def update_vehicle_service(user_id: int, vehicle_id: int, vehicle_model = VehicleModel):
     is_user_admin = select_user_is_admin(user_id)
-    if is_user_admin[0]["is_admin"]:
-        return update_vehicle_by_id(vehicle_id, vehicle_model)
+    if is_user_admin:
+        if is_user_admin[0]["is_admin"]:
+            return update_vehicle_by_id(vehicle_id, vehicle_model)
+        else:
+            return { "message" : "No such user"}
     return { "message" : "Access Denied"}
 
 def delete_vehicle_service(user_id: int, vehicle_id: int):
     is_user_admin = select_user_is_admin(user_id)
-    if is_user_admin[0]["is_admin"]:
-        return delete_vehicle_by_id(vehicle_id)
+    if is_user_admin:
+        if is_user_admin[0]["is_admin"]:
+            return delete_vehicle_by_id(vehicle_id)
+    else:
+        return { "message" : "No such user"}
     return { "message" : "Access Denied"}
